Fix SigInfo implementation
The struct needed to be aligned on eight bytes, but the preamble only
contains three 4-byte values. Correctly aligned, this means a space
between that last value and the union containing the fields of four
bytes, but Rust has no alignment syntax so that space wasn't accounted
for, meaning the first four bytes of the signal fields were bogus, and
all the actual data was shifted four bytes beyond where it should have
been.

Fix this by adding some padding to deal with alignment.