Multi-line wrapping is a go
TODO: maybe rework Controls to work primarily on a line-basis

If the above is done, it would allow for one representation of the data
written out, and instead of caring about having the rows correctly
sized, we could just calculate the offset on the next row on the fly.

However, it's not very often people resize their terminals a lot, so it
might make more sense to base our implementation on the row size rather
than the line size, so that typing is faster and the such.