Partial conversion to new io/path/process
Also partial implementation of vectrack

TODO: finish that implementation

Of note, is that we can't use the new process in command.rs yet. The
issue is there is currently no way to do InheritFd, which is a must for
piping processes to one another. Once that lands, command.rs will transition.